requirements.txt
# Kinetic PDF Processor - Dependencies
# Install met: pip install -r requirements.txt

# Core PDF verwerking
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# Beeldverbetering (optioneel maar aanbevolen)
opencv-python>=4.8.0
numpy>=1.24.0

# Web interface
streamlit>=1.28.0

# Let op: Tesseract en Poppler moeten apart geïnstalleerd worden!
# macOS:   brew install tesseract tesseract-lang poppler
# Ubuntu:  sudo apt-get install tesseract-ocr tesseract-ocr-nld poppler-utils
